Bug fix to prevent unnecessary Auth Reindex
Problem:  when selecting a collection or community to reindex, Fez is
doing an Auth Reindex of all files in that community/collection - even
when there are no authorisation changes for that collection.   There
were two problems: 1) the function setIndexAuthBGP is called within
setIndexAuthBGP for recursion, but $topcall is not included in the call.
Since the default value is true, the function always thinks it is the
top call.  Topcall is now included in the recursive call, set to false.
2) the part of the script that was comparing the fez_auth_index2 table
to the Fedora values was flawed because it was comparing authi_role as
an integer (from the database) and a text string from Fedora.  The
database query has been altered to fetch the text string.